nthDigit and nthDigitBack added
One will find the nth digit from the left to the right and one will find the nth digit from right to left

I added a try/catch exception in the event that you ask it for the nth digit out of scope for that number. I am sure there are better ways to do it, cleanly. But this is working now, so I will iterate.